backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI 框架
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langgraph==0.0.20

# 任务队列
celery==5.3.4
redis==5.0.1

# 文件处理
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# 工具
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
